compositor: Send mouse move event to ensure window focus.
sendMouseWheelEvent() doesn't ensure that the window gets focused leading to wheel events never arriving in applications until the window is hovered by means of a sendMouseMoveEvent() event.
This works well for desktop applications but not for mobile applications where one might want to use a rotary input the moment a window is opened.
This sends a sendMouseMoveEvent() call such that the ensureEntered() is called which calls the send_enter() which appears to be needed in order for wheel events to arrive (it informs the Wayland client).

Signed-off-by: Darrel Griët <dgriet@gmail.com>